Refactor: GitHubService #30
