+++ .config/dotnet-tools.json
@@ -3,7 +3,7 @@
   "isRoot": true,
   "tools": {
     "dotnet-ef": {
-      "version": "7.0.1",
+      "version": "7.0.2",
       "commands": [
         "dotnet-ef"
       ]
