+++ global.json
@@ -1,5 +1,6 @@
 {
   "sdk": {
-    "version": "7.0.100"
+    "version": "7.0.0",
+    "rollForward": "latestFeature"
   }
 }
