+++ src/AppleTv.Movie.Price.Tracker.App/Constants.cs
@@ -5,4 +5,7 @@ public class Constants
     public const string DEFAULT_CORS_POLICY = "default-cors-policy";
 
     public const string RESPONSE_MEDIA_TYPE = "application/json";
+
+    public readonly static string[] ISSUE_LABELS = new string[] { "bug", "help wanted" };
 }
+
