+++ src/AppleTv.Movie.Price.Tracker.App/appsettings.json
@@ -35,7 +35,9 @@
   "GitHub": {
     "Owner": "",
     "Repo": "",
-    "AccessToken": ""
+    "AccessToken": "",
+    "CreateNewIssueAlways": false,
+    "ReopenIfClosedOneExists": true
   },
   "Logging": {
     "LogLevel": {
