--- src/AppleTv.Movie.Price.Tracker.Services/Exceptions/GitHubError.cs
@@ -1,11 +1,0 @@
-using System.Text.Json.Serialization;
-
-namespace AppleTv.Movie.Price.Tracker.Services.Exceptions;
-
-public class GitHubError
-{
-    public string Message { get; set; }
-
-    [JsonPropertyName("documentation_url")]
-    public string DocumentationUrl { get; set; }
-}
