--- src/AppleTv.Movie.Price.Tracker.Services/GitHubOptions.cs
@@ -1,14 +1,0 @@
-namespace AppleTv.Movie.Price.Tracker.Services;
-
-public class GitHubOptions
-{
-    public const string Name = "GitHub";
-
-    public string Owner { get; set; } = "";
-
-    public string Repo { get; set; } = "";
-
-    public string AccessToken { get; set; } = "";
-}
-
-
