--- src/AppleTv.Movie.Price.Tracker.Services/Models/GitHub/GetIssueAssginees.cs
@@ -1,7 +1,0 @@
-namespace AppleTv.Movie.Price.Tracker.Services.Models.GitHUb;
-
-public class GetIssueAssginees
-{
-    public const string All = "*";
-    public const string Excerpt = "none";
-}
