--- src/AppleTv.Movie.Price.Tracker.Services/Models/GitHub/GetIssueMilestones.cs
@@ -1,7 +1,0 @@
-namespace AppleTv.Movie.Price.Tracker.Services.Models.GitHUb;
-
-public class GetIssueMilestones
-{
-    public const string All = "*";
-    public const string Excerpt = "none";
-}
