--- src/AppleTv.Movie.Price.Tracker.Services/Models/GitHub/GetIssueSortDirection.cs
@@ -1,7 +1,0 @@
-namespace AppleTv.Movie.Price.Tracker.Services.Models.GitHUb;
-
-public class GetIssueSortDirection
-{
-    public const string Ascending = "asc";
-    public const string Descending = "desc";
-}
