--- src/AppleTv.Movie.Price.Tracker.Services/Models/GitHub/GetIssueSorts.cs
@@ -1,8 +1,0 @@
-namespace AppleTv.Movie.Price.Tracker.Services.Models.GitHUb;
-
-public class GetIssueSorts
-{
-    public const string Created = "created";
-    public const string Updated = "updated";
-    public const string Comments = "comments";
-}
