--- src/AppleTv.Movie.Price.Tracker.Services/Models/GitHub/GetIssueStates.cs
@@ -1,9 +1,0 @@
-namespace AppleTv.Movie.Price.Tracker.Services.Models.GitHUb;
-
-public class GetIssueStates
-{
-    public const string Open = "open";
-    public const string Closed = "closed";
-
-    public const string All = "all";
-}
