--- src/AppleTv.Movie.Price.Tracker.Services/Models/GitHub/IIssueNumberModel.cs
@@ -1,6 +1,0 @@
-namespace AppleTv.Movie.Price.Tracker.Services.Models.GitHUb;
-
-public interface IIssueNumberModel
-{
-    public long IssueNumber { get; }
-}
