--- src/AppleTv.Movie.Price.Tracker.Services/Models/GitHub/IOwnerRepoModel.cs
@@ -1,8 +1,0 @@
-namespace AppleTv.Movie.Price.Tracker.Services.Models.GitHUb;
-
-public interface IOwnerRepoModel
-{
-    string Owner { get; }
-
-    string Repo { get; }
-}
