--- src/AppleTv.Movie.Price.Tracker.Services/Models/GitHub/IssueOrPrs.cs
@@ -1,7 +1,0 @@
-namespace AppleTv.Movie.Price.Tracker.Services.Models.GitHUb;
-
-public class IssueOrPrs
-{
-    public const string Issue = "issue";
-    public const string Pr = "pr";
-}
