--- src/AppleTv.Movie.Price.Tracker.Services/Models/GitHub/IssueStateReasons.cs
@@ -1,9 +1,0 @@
-namespace AppleTv.Movie.Price.Tracker.Services.Models.GitHUb;
-
-public class IssueStateReasons
-{
-
-    public const string Completed = "completed";
-    public const string NotPlanned = "not_planned";
-    public const string Reopened = "reopened";
-}
