--- src/AppleTv.Movie.Price.Tracker.Services/Models/GitHub/IssueStates.cs
@@ -1,7 +1,0 @@
-namespace AppleTv.Movie.Price.Tracker.Services.Models.GitHUb;
-
-public class IssueStates
-{
-    public const string Open = "open";
-    public const string Closed = "closed";
-}
