--- src/AppleTv.Movie.Price.Tracker.Services/Models/GitHub/SearchSorts.cs
@@ -1,10 +1,0 @@
-namespace AppleTv.Movie.Price.Tracker.Services.Models.GitHUb;
-
-/// <summary>
-/// https://docs.github.com/en/rest/search?apiVersion=2022-11-28#search-issues-and-pull-requests
-/// </summary>
-public class SearchSorts : GetIssueSorts
-{
-    public const string Reactions = "reactions";
-}
-
